phone-checker-bot/requirements.txt
python-telegram-bot==21.4
twilio==9.2.3
duckduckgo-search==5.3.0
openai==1.43.0
oracledb==2.2.1
python-dotenv==1.0.1
